Remove circular dependency between modules.
