+++ pg/pg.d.ts
@@ -4,13 +4,10 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
 /// <reference path="../node/node.d.ts" />
-/// <reference path="../pg-pool/pg-pool.d.ts" />
 
 declare module "pg" {
     import events = require("events");
     import stream = require("stream");
-
-    export {Pool, PoolConfig} from "pg-pool";
 
     export function connect(connection: string, callback: (err: Error, client: Client, done: (err?: any) => void) => void): void;
     export function connect(config: ClientConfig, callback: (err: Error, client: Client, done: (err?: any) => void) => void): void;
@@ -36,6 +33,16 @@ declare module "pg" {
         ssl?: boolean;
     }
 
+    export interface PoolConfig extends ClientConfig {
+      // properties from module 'node-pool'
+      max?: number;
+      min?: number;
+      refreshIdle?: boolean;
+      idleTimeoutMillis?: number;
+      reapIntervalMillis?: number;
+      returnToHead?: boolean;
+    }
+
     export interface QueryConfig {
         name?: string;
         text: string;
@@ -51,6 +58,32 @@ declare module "pg" {
 
     export interface ResultBuilder extends QueryResult {
         addRow(row: any): void;
+    }
+
+    export class Pool extends events.EventEmitter {
+
+        constructor();
+
+        // `new Pool('pg://user@localhost/mydb')` is not allowed.
+        // But it passes type check because of issue:
+        // https://github.com/Microsoft/TypeScript/issues/7485
+        constructor(config: PoolConfig);
+
+        connect(): Promise<Client>;
+        connect(callback: (err: Error, client: Client, done: () => void) => void): void;
+
+        end(): Promise<void>;
+
+        query(queryText: string): Promise<QueryResult>;
+        query(queryText: string, values: any[]): Promise<QueryResult>;
+
+        query(queryText: string, callback: (err: Error, result: QueryResult) => void): void;
+        query(queryText: string, values: any[], callback: (err: Error, result: QueryResult) => void): void;
+
+        public on(event: "error", listener: (err: Error, client: Client) => void): this;
+        public on(event: "connect", listener: (client: Client) => void): this;
+        public on(event: "acquire", listener: (client: Client) => void): this;
+        public on(event: string, listener: Function): this;
     }
 
     export class Client extends events.EventEmitter {
